This is amazing.  Mobile browsers have a virtual viewport that is larger than the devices viewport and content is rendered onto this and zoomed in to fit.
This was causing all sorts of issues on the document page however you
can set the virtual viewport to equal the devices size using this meta
tag.

Solves the problem with the size of the buttons and the reponsive design
not kicking in *and* the problem with the right panel not sliding to the
correct place.  neat.