Add similarity.{js,css} from cudl-viewer-bubbles @ 0.0.1
Was going to use git-subtree to pull in just the dist dir, but it
doesn't let you specify a subtree to pull in, so I'll have to include
these manually for the time being.

My plan is to use webpack to build the cudl-viewer UI, which will let
us reference the external similarity code without directly including it
and updating on every change.