Remove unused tablib definition in web.xml
As far as I can see this is not used, and the location it references
does not exist. Spring forms tags are used in feedback.jsp, but
they use the taglib URI http://www.springframework.org/tags/form.
The relative URI /spring in the definition I'm removing is resolved
against the web.xml's location URI, and although I'm not sure what
that is (file:// URI?) it's certainly not going to be
http://www.springframework.org, and even if it was, the path /form
would then resolve to http://www.springframework.org/form not
/tags/form.

I'm removing this because it triggers an error when using the
tomcat7-maven-plugin, and it doesn't seem to have any purpose.