Change how ui resources are bundled
The UI resources are now included as a jar archive, with resources in
META-INF/resources/ui. The servlet 3.0 spec allows for files in
META-INF/resources to be served as if they were in the .war root.

This allows the maven tomcat7 plugin to find the resources when using
the run goal. With the previous strategy of copying the resources
into target/, the tomcat7 plugin wouldn't find the ui resources, as
it'd just use src/webapp as the war directory, ignoring stuff in
target/.

This should also make it easier to run the viewer without using maven
(e.g. via an IDE).