Add support for variable recall search
The advanced search page now takes a query parameter: tagging
If set to a truthy value (1, true etc) the page will show a
slider allowing users to adjust the level of recall in the search.

This is implemented using tag data data mined from secondary
literature, as well as crowdsourced annotation data. As the slider
is moved right, the recall increases as additional sources of data
are included in the search, but precision potentially decreases,
as the additional data is quite possibly less reliable/accurate.