Merge logincontroller servlet into main dispatcher
Had to keep the raven servlet separate, as it depends on having fixed
servlet prefix path for both the login an callback URLs. If it's
mounted on the root servlet then it doesn't work. This should be
resolved in a future refactoring job to properly integrate Raven (and
the rest of CUDL's auth infrastructure) with spring-security, rather
than calling spring-security methods in ad-hoc controller methods as
we do now.