Support specifying redirect URL with query param
A next=<url> query param can be specified on the login page, which
will result in a request for the given URL being saved, and used as
the redirect target after successful authentication.

Additionally the first auth entry point automatically sets the next
query param from the page that was attempted to be viewed when the
auth exception triggered a login.