Support redirecting to URLs with fragments post login
Spring security uses its SavedRequest interface to remember a request
to replay post-login. Because it's based around HttpServletRequests, it
knows nothing of URL fragments. In order to not waste the time
previously invested in sticking with the standard Spring Security
infrastructure, I've had to devise a way of saving and restoring
fragment identifiers with the SavedRequest/RequestCache machinery.

This is a fantastic example of the kind of time-wasting bullshit you
end up doing in order to use Spring Security. I wish I'd not carried on
using it, it's such a waste of space.</rant>