Use more modern tomcat7 version via mvn cargo
As detailed in [1], the maven tomcat7 plugin was using an old version
of tomcat, and the only way to upgrade it is horrible[2]. The old
version lead to a bug being missed as live was using a more modern
version of tomcat than the version used for local testing.

This replaces the tomcat7 plugin with the Maven Cargo plugin, which
allows easily specifying the version of tomcat 7.x to use. It also
supports containers other than tomcat 7.x, so smooths the path for
upgrading to tomcat 8 (and soon 9), or alternate containers, such as
Jetty.

1: https://bitbucket.org/CUDL/cudl-viewer/issues/96/
2: http://tomcat.apache.org/maven-plugin-trunk/tomcat7-maven-plugin/adjust-embedded-tomcat-version.html