Adding an alternative embedded tomcat for local testing
Adding back in the original embedded tomcat 7 as well as the new cargo
plugin version. This is because the new version does not seem to support
the live reloading of static assets, for example JSP pages. 

This means that when you are editing a JSP page you need to stop and
start this container before you see any changes, which makes development
a lot harder.  

There may be a configuration that allows this but from what I read the
cargo plugin does not currently support this, so by adding back in the
original container, develops can use this when needed and do final
testing on the cargo version (as this has a more up to date tomcat).