Fix URIs not being encoded
Annoyingly, URI and UriComponents/UriComponentsBuilder allow the
creation of invalid URI strings. For URI, calling .toString() doesn't
encode components, so query values/paths, etc can contain invalid
characters. For UriComponentsBuilder, calling .build().toUriString()
returns an unencoded String (which is therefore not a valid URI) unless
you call .build().encode().toUriString(). A shorthand for this is just
.toUriString() (without calling build()).

Specifically, this fixes an issue reported by Mel where searching for
Yaḥyā b. Manasseh would result in an error. This was because
HttpURLConnection was requesting a URL with the above search term
unescaped as a query parameter, which caused the request to be invalid.