Use Spring Boot & embedded Tomcat
I've removed all the version properties in the pom, other than the two
internal UI resource dependencies which we're likely to bump on
releases. The version properties conflict with spring boot's parent POM
by overwriting its versions, causing odd problems.

UnsupportedEncodingException is now thrown less. UriUtils used to
threaten to throw it whenever a charset name was used, but it seems to
not to do that anymore!

Need to:
- clean up configs / make use of spring boot config system