Move the equatorie src dir out of the webapp
We don't want to package it, it shouldn't actually be in this repo at
all, but that's something for another day. The maven war plugin
doesn't seem to provide a way to avoid copying it to the taret dir
during a build; which is reasonable, as you'd expect things in the
webapp dir to go into a webapp.