Fix /contributors page not being editable
The page wasn't including sec: and json: taglibs, which was causing
the <sec:authorize> tag to not work. It did used to work, so presumably
some kind of indirect taglib definition used to happen, but no longer
does. If so, other things could be broken now.

This fixes it by using the proper <cudl:context-editable-areas> tag,
which does all the auth checking. CUDL-196