Fix partial responses when JSP errors occur
Previously, if an error occurred mid-way through JSP rendering, the
client would still receive an HTTP 200, but with only part of the page
content (up to the point that the error occurred). Also the HTTP
response would be technically invalid, as the body is sent with
Transfer-Encoding: Chunked, and it wasn't finished.

We now set the JSP buffer to a value large enough to hold generated
HTML, and disable incremental flushing, which allows a proper error
response to be sent in the event of errors occurring.