Support tests using the Viewer's ApplicationContext
This slightly tweaks the bean definitions of the viewer to allow tests
to load a Spring ApplicationContext containing the Viewer's beans, so
that they can test controller methods and the logic of beans in
general.

Tests needing to use the ApplicationContext should extend
BaseCUDLApplicationContextTest.

The tests run with the "test" profile enabled; beans registered via
@Component annotations, or requiring external services at runtime
(e.g. databases) should be excluded from the "test" profile, and have
suitable default mocks defined in ParentTestConfig and ChildTestConfig.