Implement automatic mock bean resetting
Mock beans have a lifetime that bridges multiple tests, so if they're
not reset between tests then validations of method calls are not
reliable, as invocations in independent tests interfere with one
another.

This implements a method for mock beans to register themselves as
needing to be reset, and provides an API for tests to reset the beans.