Implement generation of image URLs for item pages
The assumption now is that item pages have a single property defining
the URLs of images representing the page. It looks like this:

    "IIIFImageURL": "MS-ADD-03419-000-00001"

Prevously we had two situations. Items would either specify URLs for
the DZI image server like so:

    "displayImageURL": "/content/images/MS-ADD-03419-000-00001.dzi",
    "downloadImageURL": "/content/images/MS-ADD-03419-000-00001.jpg",
    "thumbnailImageURL": "/content/images/MS-ADD-03419-000-00001_files/8/0_0.jpg",

Or specify the a IIIF image URL (from which the others would be
derived as needed):

    "IIIFImageURL": "MS-ADD-03419-000-00002.jp2",

Huw had got himself into a difficult situation by creating a fork of
the data repository with each of these image URL representations.
Because of the delay in rolling out production IIIF support, he's had
to maintain both forks longer than expected.

This change is to allow him to drop the DZI data fork, so he only has
one to maintain.

In the future when we roll out production IIIF image servers we can
stop the generation of DZI URLs that this change implements.