Fix admin controller denying users with ADMIN role
The @Secured annotation was being used with a SpEL expression, but it
only accepts a list of role names to check. I'm not sure how this used
to work, as even in Spring Security 3 this is how @Secured worked.

Anyway, @PreAuthorize does evaluate SpEL expressions, and this works
as expected.