Add feature flag for image server type
The images.server.type property can be either 'dzi-and-iiif' or 'iiif',
depending on whether CUDLs legacy image server is used (mix of
pre-generated images, DZIs and IIIF for restricted purposes) or a
purely IIIF image server is used.

This requires the commit
    "Support IIIF and pre-generated/DZI images simultaneously"
from cudl-viewer-ui to have the UI use either DZIs or IIIF for zooming
images, depending on which is available in the metadata.